refactor: 绘图 API 坐标形参由 int16_t 加宽为 int, 消除窄化转换
批次 14/15。**本批次有 ABI 变化与代码体积代价, 需实机 OLED 逐页回归 (H9)**。

背景: Clang-Tidy 的 bugprone-narrowing-conversions 在本工程报出约 90 处,
全部集中在 WouoUI_page.c 与 WouoUI_graph.c, 根因是绘图 API 用 int16_t 作
坐标形参, 而调用处传入 int / uint16_t 表达式 —— 每次调用报一条, 逐行加 cast
会制造 90 处噪音。

改动: 把坐标/尺寸类形参统一加宽为 int, 涉及
  Core/Inc/WouoUI_graph.h  Core/Src/WouoUI_graph.c
  Core/Inc/WouoUI_page.h   Core/Src/WouoUI_page.c

**未改动**: Canvas / SlideStr / WavePage 等结构体成员 (数据布局不变)、
函数返回值类型、局部变量、强制转换。

执行细节 (踩过的坑, 记录备查):
- 裸 sed 子串替换会误伤 PULSE_MODE_NPULSE 一类同前缀宏 —— 但这里改的是
  int16_t, 无前缀冲突, 用的是"( 或 , 紧跟 int16_t + 空格"的上下文模式
- 第一版模式要求逗号后必须有空格, 漏掉了 `ss,int16_t y` 这种无空格写法,
  以及 `, int16_t ` 这种带空格写法 —— 两种都补齐后才通过编译
- 声明与定义必须同步改 (WouoUI_page.h 漏改会导致 conflicting types 编译错误)
- 顺带修正了若干已过时的 Doxygen 注释中的旧签名

验证:
- 编译零错误零告警
- 符号变化仅为 LTO 对同一批图形函数的 .constprop / .isra 特化变体重组,
  无函数丢失 (DrawASCII 的一个特化被内联, DrawSlideStr 的一个特化被拆分)
- 已核验全部改动行都落在函数形参列表及其文档注释上, 无误伤
- 已在两个 MPL-2.0 文件的版权头中登记修改记录

**代价 (需权衡)**: FLASH 136028 → 137668, **+1640 字节 (+1.2%)**。
原因: int16_t 形参让 GCC 能证明取值范围并消除部分冗余边界检查; 加宽后这些
检查保留, 且 -funroll-loops 下寄存器压力上升。512KB Flash 当前占用 26.26%,
空间充裕, 故接受。**若不接受, 回退本提交即可, 改用 .clang-tidy 抑制
bugprone-narrowing-conversions 作为替代方案。**

**待实机验证 (计划 R4/H9, 不可省)**: OLED 全部页面逐页目视比对
(主页面 / 7 个模式页 / 通道·极性·互补对选择页 / Setting / About),
须与改动前像素级一致。

+++ Core/Inc/WouoUI_graph.h
@@ -94,25 +94,25 @@ void WouoUI_BuffSend(void);
     void WouoUI_GraphSetDynamicBuff(ScreenBuff* buff);
 #endif
 
-int16_t WouoUI_CanvasDrawASCII(Canvas *canvas, int16_t x, int16_t y, sFONT font, char c);
-void WouoUI_CanvasDrawStr(Canvas *canvas, int16_t x, int16_t y, sFONT font, uint8_t *str);
-void WouoUI_CanvasDrawSlideStr(SlideStr* ss, int16_t y, sFONT font);
+int16_t WouoUI_CanvasDrawASCII(Canvas *canvas, int x, int y, sFONT font, char c);
+void WouoUI_CanvasDrawStr(Canvas *canvas, int x, int y, sFONT font, uint8_t *str);
+void WouoUI_CanvasDrawSlideStr(SlideStr* ss, int y, sFONT font);
 void WouoUI_CanvasSlideStrReset(SlideStr* ss);
-void WouoUI_CanvasDrawStrWithNewline(Canvas *canvas, int16_t x, int16_t y, sFONT font, uint8_t *str, uint8_t lineSpacing);
-void WouoUI_CanvasDrawStrAutoNewline(Canvas *canvas, int16_t x, int16_t y, sFONT font, uint8_t *str);
+void WouoUI_CanvasDrawStrWithNewline(Canvas *canvas, int x, int y, sFONT font, uint8_t *str, uint8_t lineSpacing);
+void WouoUI_CanvasDrawStrAutoNewline(Canvas *canvas, int x, int y, sFONT font, uint8_t *str);
 uint16_t WouoUI_GetStrWidth(const char *str, sFONT font);
 uint16_t WouoUI_GetStrHeight(const char *str, sFONT font);
-uint16_t WouoUI_GetStrHeightAutoNewLine(int16_t canvas_w, const char* str, sFONT font);
-void WouoUI_CanvasDrawLine_V(Canvas *canvas, int16_t x, int16_t y_start, int16_t y_end);
-void WouoUI_CanvasDrawDashedLine_V(Canvas *canvas, int16_t x, int16_t y_start, int16_t y_end, uint8_t DashedStyle, uint8_t Unit_Length);
-void WouoUI_CanvasDrawLine_H(Canvas *canvas, int16_t x_start, int16_t x_end, int16_t y);
-void WouoUI_GraphDrawDashedLine_H(Canvas *canvas, int16_t x_start, int16_t x_end, int16_t y, uint8_t DashedStyle, uint8_t Unit_Length);
-void WouoUI_CanvasDrawRBox(Canvas *canvas, int16_t x_start, int16_t y_start, int16_t width, int16_t height, uint8_t r);
-void WouoUI_CanvasDrawRBoxEmpty(Canvas *canvas, int16_t x_start, int16_t y_start, int16_t width, int16_t height, uint8_t r);
-void WouoUI_CanvasDrawBoxRightAngle(Canvas *canvas, int16_t x_start, int16_t y_start, int16_t width, int16_t height, uint8_t r);
-void WouoUI_CanvasDrawBMP(Canvas *canvas, int16_t x, int16_t y, int16_t width, int16_t height, const uint8_t *BMP, uint8_t color);
-void WouoUI_CanvasDrawPoint(Canvas *canvas, int16_t x, int16_t y);
-void WouoUI_CanvasDrawLine(Canvas *canvas, int16_t x1, int16_t y1, int16_t x2, int16_t y2);
+uint16_t WouoUI_GetStrHeightAutoNewLine(int canvas_w, const char* str, sFONT font);
+void WouoUI_CanvasDrawLine_V(Canvas *canvas, int x, int y_start, int y_end);
+void WouoUI_CanvasDrawDashedLine_V(Canvas *canvas, int x, int y_start, int y_end, uint8_t DashedStyle, uint8_t Unit_Length);
+void WouoUI_CanvasDrawLine_H(Canvas *canvas, int x_start, int x_end, int y);
+void WouoUI_GraphDrawDashedLine_H(Canvas *canvas, int x_start, int x_end, int y, uint8_t DashedStyle, uint8_t Unit_Length);
+void WouoUI_CanvasDrawRBox(Canvas *canvas, int x_start, int y_start, int width, int height, uint8_t r);
+void WouoUI_CanvasDrawRBoxEmpty(Canvas *canvas, int x_start, int y_start, int width, int height, uint8_t r);
+void WouoUI_CanvasDrawBoxRightAngle(Canvas *canvas, int x_start, int y_start, int width, int height, uint8_t r);
+void WouoUI_CanvasDrawBMP(Canvas *canvas, int x, int y, int width, int height, const uint8_t *BMP, uint8_t color);
+void WouoUI_CanvasDrawPoint(Canvas *canvas, int x, int y);
+void WouoUI_CanvasDrawLine(Canvas *canvas, int x1, int y1, int x2, int y2);
 void WouoUI_BuffAllBlur(BLUR_DEGREE blur);
 
 #ifdef __cplusplus
